+++ NEWS.md
@@ -1,6 +1,7 @@
 # v. 0.1.5
 
   * `mpspline_datchk()` now removes 0-thickness horizons (where upper and lower depths are identical)
+  * `mpspline2_datchk()` now removes horizons where upper depth is greater than lower depth
   * `mpspline_datchk()` now returns messages about each edit made to the input data
 
 # v. 0.1.4